requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
pyjwt
requests
web3
py-solc-x